Add specialized `vec_joint_proxy_order()` for list-column support
Lists have an order-proxy that uses first appearance order. This means that you can't compare the results of the order-proxy of `x` and the order-proxy of `y` taken separately. You have to take them jointly to get correct first appearance values. I've hardcoded this path to just intercept `vec_proxy_order.list()`, as I don't think this is going going to affect any other type. If this changes, this may have to be rethought.