Realign complement with link function
This means `[a, b)` and `[b, c)` no longer generate the empty set of `[b, b)` as their complement. Instead, it generates a zero row result. This aligns better with linking, because linking those two intervals first would give `[a, c)` by default, and taking the complement of that would return the zero row result.