Treat empty intervals as if they provide no information at all
We now remove them completely up front, which solves a number of implementation headaches. We can now assume that `start < end` unconditionally.

This also ensures that an empty interval in `y` of `interval_setdiff()` won't affect the result, which was the original reason for this change.

Note that IRanges provides `drop.empty.ranges` and defaults it to false, but then sets it to true manually everywhere it is used in the package. The spirit of this commit would be equivalent to setting this argument to true unconditionally, with the rationale that empty intervals never add information. IRanges also defines a "normal" range as one without empty ranges and claims that `reduce()` gives this, but it doesn't by default since `drop.empty.ranges` defaults to false.