Ensure location determination works with ignored empty ranges
If the empty ranges are removed, we need to do some additional work to get the locations right, as they no longer match up to the original locations anymore. This can be easily accomplished by retaining the results of `which(non-empty)` and using that as a map back to the original locations.